Relax Security when visiting H2 Console
Not fit for a production configuration, but is a fitting solution
in a development environment. A further improvement would be to
annotate this `SecurityConfig` class with `@Profile("dev")` so that
we only expose H2 Console access while the Spring Profile `dev` is
active, while creating `ProdSecurityConfig` class, having the same
content as previous, but annotated with `@Profile("prod")`.

* Added the overloaded `configure(WebSecurity web)` method to
`SecurityConfig`, which ignores security constraints when accessing
the H2 console.